commit to do pull

+++ src/main/java/academy/softserve/eschool/controller/ScheduleController.java
@@ -3,6 +3,7 @@ package academy.softserve.eschool.controller;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 
+import academy.softserve.eschool.repository.MarkRepository;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -37,6 +38,8 @@ public class ScheduleController {
     ScheduleServiceImpl scheduleService;
     @NonNull
     LessonRepository lessonRepository;
+    @NonNull
+    MarkRepository markRepository;
 
     @ApiOperation(value = "Creates a schedule for a class")
     @ApiResponses(
@@ -50,12 +53,17 @@ public class ScheduleController {
     @PostMapping("/classes/{classId}/schedule")
     //todo bk ++
     public GeneralResponseWrapper<ScheduleDTO> postSchedule(
-            @ApiParam(value = "schedule object", required = true) @RequestBody ScheduleDTO scheduleDTO)
-    {
+            @ApiParam(value = "schedule object", required = true) @RequestBody ScheduleDTO scheduleDTO){
         //todo bk
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
         LocalDate startDate = LocalDate.of(scheduleDTO.getStartOfSemester().getYear(), scheduleDTO.getStartOfSemester().getMonth(), scheduleDTO.getStartOfSemester().getDayOfMonth());
         LocalDate endDate = LocalDate.of(scheduleDTO.getEndOfSemester().getYear(), scheduleDTO.getEndOfSemester().getMonth(), scheduleDTO.getEndOfSemester().getDayOfMonth());
+        /*
+        if(markRepository.getCountOfMarksByDateBounds((startDate).format(formatter), (endDate).format(formatter)) > 0 )
+        {
+            return new GeneralResponseWrapper<>(new Status(409, "Conflict"), null);
+        }
+        */
         lessonRepository.deleteScheduleByBounds((startDate).format(formatter), (endDate).format(formatter),
                 scheduleDTO.getClassName().getId());
         scheduleService.saveSchedule(scheduleDTO);
